requirements.txt
bs4
requests
colorama
darkdetect
mammoth
openpyxl
python-pptx
pdfminer.six
Pillow
pymupdf
python-docx
google-genai
pdf2image
